Update webin-cli.md
Added many missing values for both these metadata lists. Permitted values for platform and Permitted values for instrument. These values were present in the SRA_experiment.xsd but were not on this markdown. Moved some terms under a deprecated header for both platform and instrument sections. Researched and created definitions that were missing for many of the platforms; not perfect, but far better than none.