Update webin-cli.md with library_selection
Updated library_selection with these 4 terms that are in the XSD, but not yet in the readme. Used the definition from the XSD,
snRNA-seq
NOMe-Seq
ChM-Seq
GBS
N.B. snRNA-seq(single nucleus) may be problematic due to the current definitions of ncRNA-Seq which also as  a snRNA-seq its definition, but that is probably meaning (small nuclear). Suggest that the definition of ncRNA-Seq is tweaked. Does that need to be cross-INSDC agreed?